fix: strengthen weak test assertions and split oversized test files (#579)
Weak/tautological assertions: SocialIcon now asserts the correct
brand-icon path per platform instead of "an svg exists"; Select's
change-event test asserts the propagated value; WaveformSeekBar's
mount test asserts bars are actually drawn from fetched waveform data.

Oversized files: collection.test.ts split into library/search/
navigation/layout/miniplayer concerns; theme.test.ts split into
colorAccessibility.test.ts (WCAG/contrast data checks) and
themeStore.test.ts (color-extraction utilities + ThemeStore behavior).
