--- src/lib/stores/collection.test.ts
@@ -1,617 +1,0 @@
-import { describe, it, expect, beforeEach, vi } from "vitest";
-import { invoke } from "@tauri-apps/api/core";
-import { listen } from "@tauri-apps/api/event";
-import type { Song, AlbumItem, ArtistItem } from "../types";
-
-const windowGeometryCallbacks = vi.hoisted(() => [] as Array<() => void>);
-
-vi.mock("@tauri-apps/api/window", () => ({
-  getCurrentWindow: vi.fn(() => ({
-    onResized: vi.fn((cb: () => void) => {
-      windowGeometryCallbacks.push(cb);
-      return Promise.resolve(() => {});
-    }),
-    onMoved: vi.fn((cb: () => void) => {
-      windowGeometryCallbacks.push(cb);
-      return Promise.resolve(() => {});
-    }),
-  })),
-}));
-
-import { collectionStore } from "./collection.svelte";
-
-describe("CollectionStore", () => {
-  let eventCallbacks: Record<string, Function> = {};
-
-  beforeEach(() => {
-    vi.clearAllMocks();
-    eventCallbacks = {};
-    // windowGeometryCallbacks is populated once, when the collectionStore
-    // singleton's constructor calls initWindowGeometryTracking() at module
-    // import time — it must not be cleared per-test, there's no later call
-    // that would repopulate it.
-
-    vi.mocked(listen).mockImplementation(async (event: string, callback: any) => {
-      eventCallbacks[event] = callback;
-      return () => {};
-    });
-
-    vi.mocked(invoke).mockImplementation(async (cmd: string, args?: any) => {
-      switch (cmd) {
-        case "geometry_capture_supported":
-          return true;
-        case "get_directories":
-          return [{ id: 1, path: "/music/rock", created_at: "2026-01-01" }];
-        case "get_library_stats":
-          return {
-            total_songs: 10,
-            total_artists: 2,
-            total_albums: 2,
-            total_duration_nanosec: 30000000000,
-            total_filesize_bytes: 50000000
-          };
-        case "get_library_snapshot":
-          return {
-            songs: [
-              { id: 1, title: "Rock Track 1", artist: "Rock Band", album: "Rock Album", filetype: "MP3" },
-              { id: 2, title: "Jazz Track 1", artist: "Jazz Quartet", album: "Jazz Album", filetype: "FLAC" },
-              { id: 3, title: "Vorbis Track", artist: "Indie Group", album: "Indie Album", filetype: "OGG_VORBIS" }
-            ],
-            albums: [
-              { album: "Rock Album", artist: "Rock Band", song_count: 5, year: 2020 },
-              { album: "Jazz Album", artist: "Jazz Quartet", song_count: 5, year: 2021 }
-            ],
-            artists: [
-              { name: "Rock Band", album_count: 1, song_count: 5 },
-              { name: "Jazz Quartet", album_count: 1, song_count: 5 }
-            ]
-          };
-        case "get_all_app_settings":
-          return {};
-        case "search_songs":
-          if (args?.query === "Rock") {
-            return [{ id: 1, title: "Rock Track 1", artist: "Rock Band", album: "Rock Album", filetype: "MP3" }];
-          }
-          return [];
-        default:
-          return null;
-      }
-    });
-  });
-
-  it("refreshes directories, stats, and library upon refresh calls", async () => {
-    await collectionStore.refreshDirectories();
-    expect(collectionStore.directories).toHaveLength(1);
-    expect(collectionStore.directories[0].path).toBe("/music/rock");
-
-    await collectionStore.refreshStats();
-    expect(collectionStore.stats.total_songs).toBe(10);
-
-    await collectionStore.refreshLibrary();
-    expect(collectionStore.songs).toHaveLength(3);
-    expect(collectionStore.albums).toHaveLength(2);
-    expect(collectionStore.artists).toHaveLength(2);
-  });
-
-  it("invokes backend on addDirectory and removeDirectory", async () => {
-    await collectionStore.addDirectory("/music/pop");
-    expect(invoke).toHaveBeenCalledWith("add_directory", { path: "/music/pop" });
-    expect(invoke).toHaveBeenCalledWith("get_directories");
-
-    await collectionStore.removeDirectory("/music/pop");
-    expect(invoke).toHaveBeenCalledWith("remove_directory", { path: "/music/pop" });
-  });
-
-  it("handles directory scanning and scan-progress event with force option", async () => {
-    vi.mocked(invoke).mockResolvedValueOnce(undefined as any);
-    await collectionStore.startScan(true);
-    expect(collectionStore.isScanning).toBe(true);
-    expect(invoke).toHaveBeenCalledWith("scan_directories", { force: true });
-
-    if (eventCallbacks["scan-progress"]) {
-      eventCallbacks["scan-progress"]({
-        payload: { phase: "reading_tags", current_path: "song.mp3", scanned: 5, total: 10 }
-      });
-      expect(collectionStore.scanProgress?.scanned).toBe(5);
-      expect(collectionStore.isScanning).toBe(true);
-
-      eventCallbacks["scan-progress"]({
-        payload: { phase: "done", current_path: "", scanned: 10, total: 10 }
-      });
-      expect(collectionStore.isScanning).toBe(false);
-      expect(collectionStore.lastScanTime).not.toBeNull();
-    }
-  });
-
-  it("handles pruneMissing songs call", async () => {
-    vi.mocked(invoke).mockResolvedValueOnce({ deleted_songs: 3, removed_folders: 2, merged_duplicates: 1 } as any);
-    const result = await collectionStore.pruneMissing();
-    expect(invoke).toHaveBeenCalledWith("prune_missing_songs");
-    expect(result).toEqual({ deletedSongs: 3, removedFolders: 2, mergedDuplicates: 1 });
-  });
-
-  it("toggles and persists watchFoldersRealtime and scanOnStartup settings", async () => {
-    await collectionStore.setWatchFoldersRealtime(false);
-    expect(collectionStore.watchFoldersRealtime).toBe(false);
-    expect(invoke).toHaveBeenCalledWith("set_app_setting", { key: "watch_folders_realtime", value: "false" });
-
-    await collectionStore.setScanOnStartup(true);
-    expect(collectionStore.scanOnStartup).toBe(true);
-    expect(invoke).toHaveBeenCalledWith("set_app_setting", { key: "scan_on_startup", value: "true" });
-  });
-
-  it("executes FTS search and updates search results and loading state", async () => {
-    await collectionStore.search("Rock");
-    expect(invoke).toHaveBeenCalledWith("search_songs", { query: "Rock", limit: 500 });
-    expect(collectionStore.searchQuery).toBe("Rock");
-    expect(collectionStore.searchResults).toHaveLength(1);
-    expect(collectionStore.searchResults[0].title).toBe("Rock Track 1");
-
-    await collectionStore.search("");
-    expect(collectionStore.searchResults).toHaveLength(0);
-  });
-
-  it("filters albums and artists by search query", () => {
-    collectionStore.albums = [
-      { album: "Dark Side", artist: "Pink Floyd" } as AlbumItem,
-      { album: "Abbey Road", artist: "The Beatles" } as AlbumItem
-    ];
-    collectionStore.artists = [
-      { name: "Pink Floyd" } as ArtistItem,
-      { name: "The Beatles" } as ArtistItem
-    ];
-
-    collectionStore.searchQuery = "beatles";
-    expect(collectionStore.filteredAlbums).toHaveLength(1);
-    expect(collectionStore.filteredAlbums[0].album).toBe("Abbey Road");
-
-    expect(collectionStore.filteredArtists).toHaveLength(1);
-    expect(collectionStore.filteredArtists[0].name).toBe("The Beatles");
-  });
-
-  it("filters artists by tags using artist-tag prefix and general search", () => {
-    collectionStore.artists = [
-      { name: "Shania Twain" } as ArtistItem,
-      { name: "The Beatles" } as ArtistItem,
-      { name: "Celine Dion" } as ArtistItem,
-    ];
-    collectionStore.artistProfiles = {
-      "shania twain": {
-        artist_key: "Shania Twain",
-        tags: ["country", "canadian", "pop"],
-        social_links: [],
-      },
-      "celine dion": {
-        artist_key: "Celine Dion",
-        tags: ["pop", "canadian", "ballad"],
-        social_links: [],
-      },
-      "the beatles": {
-        artist_key: "The Beatles",
-        tags: ["rock", "british", "classic"],
-        social_links: [],
-      },
-    };
-
-    // Explicit tag search
-    collectionStore.searchQuery = "artist-tag:canadian";
-    expect(collectionStore.filteredArtists).toHaveLength(2);
-    expect(collectionStore.filteredArtists.map((a) => a.name)).toEqual(["Shania Twain", "Celine Dion"]);
-
-    collectionStore.searchQuery = "tag:country";
-    expect(collectionStore.filteredArtists).toHaveLength(1);
-    expect(collectionStore.filteredArtists[0].name).toBe("Shania Twain");
-
-    // General keyword search matching tag
-    collectionStore.searchQuery = "british";
-    expect(collectionStore.filteredArtists).toHaveLength(1);
-    expect(collectionStore.filteredArtists[0].name).toBe("The Beatles");
-  });
-
-  it("filters albums by artist tags using tag prefix and general search", () => {
-    collectionStore.albums = [
-      { album: "Come On Over", artist: "Shania Twain" } as AlbumItem,
-      { album: "The Woman in Me", artist: "Shania Twain" } as AlbumItem,
-      { album: "Let's Talk About Love", artist: "Celine Dion" } as AlbumItem,
-      { album: "Abbey Road", artist: "The Beatles" } as AlbumItem,
-    ];
-    collectionStore.artistProfiles = {
-      "shania twain": {
-        artist_key: "Shania Twain",
-        tags: ["country", "canadian", "pop"],
-        social_links: [],
-      },
-      "celine dion": {
-        artist_key: "Celine Dion",
-        tags: ["pop", "canadian", "ballad"],
-        social_links: [],
-      },
-      "the beatles": {
-        artist_key: "The Beatles",
-        tags: ["rock", "british", "classic"],
-        social_links: [],
-      },
-    };
-
-    // Explicit tag search with tag: prefix
-    collectionStore.searchQuery = "tag:country";
-    expect(collectionStore.filteredAlbums).toHaveLength(2);
-    expect(collectionStore.filteredAlbums.map((a) => a.album)).toEqual([
-      "Come On Over",
-      "The Woman in Me",
-    ]);
-
-    // Explicit tag search with artist-tag: prefix
-    collectionStore.searchQuery = "artist-tag:canadian";
-    expect(collectionStore.filteredAlbums).toHaveLength(3);
-    expect(collectionStore.filteredAlbums.map((a) => a.album)).toEqual([
-      "Come On Over",
-      "The Woman in Me",
-      "Let's Talk About Love",
-    ]);
-
-    // General keyword search matching artist tag
-    collectionStore.searchQuery = "british";
-    expect(collectionStore.filteredAlbums).toHaveLength(1);
-    expect(collectionStore.filteredAlbums[0].album).toBe("Abbey Road");
-
-    // General keyword search matching album title directly
-    collectionStore.searchQuery = "over";
-    expect(collectionStore.filteredAlbums).toHaveLength(1);
-    expect(collectionStore.filteredAlbums[0].album).toBe("Come On Over");
-  });
-
-  it("handles navigation helpers viewArtist and viewAlbum and clears search terms", () => {
-    collectionStore.searchQuery = "some search";
-    collectionStore.searchResults = [{ id: 1 } as Song];
-
-    collectionStore.viewArtist("Pink Floyd");
-    expect(collectionStore.selectedArtistName).toBe("Pink Floyd");
-    expect(collectionStore.activeTab).toBe("collection");
-    expect(collectionStore.activeSubTab).toBe("artists");
-    expect(collectionStore.searchQuery).toBe("");
-    expect(collectionStore.searchResults).toHaveLength(0);
-
-    collectionStore.searchQuery = "another search";
-    collectionStore.viewAlbum("Dark Side");
-    expect(collectionStore.selectedAlbumName).toBe("Dark Side");
-    expect(collectionStore.searchQuery).toBe("");
-  });
-
-  it("persists the selected album/artist detail view to localStorage so a relaunch restores it", () => {
-    collectionStore.viewAlbum("Dark Side of the Moon");
-    expect(localStorage.getItem("navigation_selectedAlbumName")).toBe("Dark Side of the Moon");
-    expect(localStorage.getItem("navigation_selectedArtistName")).toBeNull();
-
-    collectionStore.viewArtist("Pink Floyd");
-    expect(localStorage.getItem("navigation_selectedArtistName")).toBe("Pink Floyd");
-    expect(localStorage.getItem("navigation_selectedAlbumName")).toBeNull();
-
-    collectionStore.selectedArtistName = null;
-    expect(localStorage.getItem("navigation_selectedArtistName")).toBeNull();
-  });
-
-  it("supports Back/Forward navigation through viewArtist/viewAlbum history", async () => {
-    // Flush the microtask-coalesced history record from any earlier test's navigation
-    // before establishing our own baseline entry (history is a shared singleton across tests).
-    await Promise.resolve();
-    collectionStore.selectedArtistName = null;
-    collectionStore.selectedAlbumName = null;
-    await Promise.resolve();
-
-    collectionStore.viewArtist("History Test Artist");
-    await Promise.resolve();
-
-    collectionStore.viewAlbum("History Test Album");
-    await Promise.resolve();
-
-    expect(collectionStore.selectedAlbumName).toBe("History Test Album");
-    expect(collectionStore.canGoBack).toBe(true);
-
-    collectionStore.goBack();
-    expect(collectionStore.selectedArtistName).toBe("History Test Artist");
-    expect(collectionStore.selectedAlbumName).toBeNull();
-    expect(collectionStore.canGoForward).toBe(true);
-
-    collectionStore.goForward();
-    expect(collectionStore.selectedAlbumName).toBe("History Test Album");
-    expect(collectionStore.canGoForward).toBe(false);
-  });
-
-  it("truncates forward history when navigating anew from a Back'd-into state", async () => {
-    await Promise.resolve();
-
-    collectionStore.viewArtist("Artist A");
-    await Promise.resolve();
-    collectionStore.viewArtist("Artist B");
-    await Promise.resolve();
-
-    collectionStore.goBack();
-    expect(collectionStore.selectedArtistName).toBe("Artist A");
-    expect(collectionStore.canGoForward).toBe(true);
-
-    collectionStore.viewArtist("Artist C");
-    await Promise.resolve();
-
-    expect(collectionStore.selectedArtistName).toBe("Artist C");
-    expect(collectionStore.canGoForward).toBe(false);
-
-    collectionStore.goBack();
-    expect(collectionStore.selectedArtistName).toBe("Artist A");
-  });
-
-  it("toggles and persists layout states (sidebar, right panel, immersive mode)", () => {
-    const initialSidebar = collectionStore.sidebarOpen;
-    collectionStore.toggleSidebar();
-    expect(collectionStore.sidebarOpen).toBe(!initialSidebar);
-
-    collectionStore.setSidebarWidth(300);
-    expect(collectionStore.sidebarWidth).toBe(300);
-
-    const initialRight = collectionStore.rightPanelOpen;
-    collectionStore.toggleRightPanel();
-    expect(collectionStore.rightPanelOpen).toBe(!initialRight);
-
-    collectionStore.setRightPanelWidth(320);
-    expect(collectionStore.rightPanelWidth).toBe(320);
-
-    collectionStore.immersiveMode = true;
-    collectionStore.exitImmersiveMode();
-    expect(collectionStore.immersiveMode).toBe(false);
-  });
-
-  it("derives responsive breakpoint flags from viewportWidth/viewportHeight without touching stored preferences", () => {
-    collectionStore.sidebarOpen = true;
-    collectionStore.setSidebarWidth(300);
-    collectionStore.rightPanelOpen = true;
-    collectionStore.immersiveMode = false;
-
-    // Large: nothing auto-collapsed.
-    collectionStore.viewportWidth = 1200;
-    collectionStore.viewportHeight = 800;
-    expect(collectionStore.isSidebarAutoCollapsed).toBe(false);
-    expect(collectionStore.isRightPanelAutoHidden).toBe(false);
-    expect(collectionStore.isImmersiveForced).toBe(false);
-    expect(collectionStore.effectiveImmersiveMode).toBe(false);
-    expect(collectionStore.isPlaybarOnlyMode).toBe(false);
-
-    // Medium: sidebar/right-panel auto-collapse, real preferences untouched.
-    collectionStore.viewportWidth = 800;
-    expect(collectionStore.isSidebarAutoCollapsed).toBe(true);
-    expect(collectionStore.isRightPanelAutoHidden).toBe(true);
-    expect(collectionStore.isImmersiveForced).toBe(false);
-    expect(collectionStore.sidebarWidth).toBe(300);
-    expect(collectionStore.rightPanelOpen).toBe(true);
-
-    // Small: immersive force-engages via the derived flag only.
-    collectionStore.viewportWidth = 500;
-    expect(collectionStore.isImmersiveForced).toBe(true);
-    expect(collectionStore.effectiveImmersiveMode).toBe(true);
-    expect(collectionStore.immersiveMode).toBe(false);
-
-    // Widening back out restores the large-tier flags without any stored
-    // preference having been mutated along the way.
-    collectionStore.viewportWidth = 1200;
-    expect(collectionStore.isSidebarAutoCollapsed).toBe(false);
-    expect(collectionStore.isRightPanelAutoHidden).toBe(false);
-    expect(collectionStore.effectiveImmersiveMode).toBe(false);
-
-    // Height is independent of width.
-    collectionStore.viewportHeight = 250;
-    expect(collectionStore.isPlaybarOnlyMode).toBe(true);
-    collectionStore.viewportHeight = 800;
-    expect(collectionStore.isPlaybarOnlyMode).toBe(false);
-  });
-
-  it("toggles sidebar compact state between 64px and expanded width", () => {
-    collectionStore.sidebarOpen = true;
-    collectionStore.setSidebarWidth(256);
-    expect(collectionStore.sidebarWidth).toBe(256);
-
-    // Toggle compact -> collapses to 64px
-    collectionStore.toggleSidebarCompact();
-    expect(collectionStore.sidebarWidth).toBe(64);
-
-    // Toggle compact again -> expands to 256px
-    collectionStore.toggleSidebarCompact();
-    expect(collectionStore.sidebarWidth).toBe(256);
-
-    // If sidebar was hidden, toggleSidebarCompact opens and expands it
-    collectionStore.sidebarOpen = false;
-    collectionStore.sidebarWidth = 64;
-    collectionStore.toggleSidebarCompact();
-    expect(collectionStore.sidebarOpen).toBe(true);
-    expect(collectionStore.sidebarWidth).toBe(256);
-  });
-
-  it("sends remembered size/position as the toggle target when geometry capture is supported, ignoring any command return value", async () => {
-    collectionStore.setMiniplayerGeometry(310, 370, 20, 30);
-    collectionStore.setSavedWindowGeometry(1400, 900, 50, 60);
-
-    vi.mocked(invoke).mockImplementation(async (cmd: string) => {
-      if (cmd === "geometry_capture_supported") return true;
-      return { width: 9999, height: 9999 };
-    });
-
-    await collectionStore.enterMiniplayerMode();
-    expect(invoke).toHaveBeenCalledWith("enter_miniplayer_mode", { width: 310, height: 370, x: 20, y: 30 });
-    expect(collectionStore.isMiniplayer).toBe(true);
-    expect(collectionStore.miniplayerWidth).toBe(310);
-
-    await collectionStore.exitMiniplayerMode();
-    expect(invoke).toHaveBeenCalledWith("exit_miniplayer_mode", { width: 1400, height: 900, x: 50, y: 60 });
-    expect(collectionStore.isMiniplayer).toBe(false);
-    expect(collectionStore.savedWindowWidth).toBe(1400);
-  });
-
-  it("sends no geometry at all when geometry capture is unsupported, letting the backend use its fixed defaults", async () => {
-    collectionStore.setMiniplayerGeometry(310, 370, 20, 30);
-
-    vi.mocked(invoke).mockImplementation(async (cmd: string) => {
-      if (cmd === "geometry_capture_supported") return false;
-      return null;
-    });
-
-    await collectionStore.enterMiniplayerMode();
-    expect(invoke).toHaveBeenCalledWith("enter_miniplayer_mode", {});
-    expect(collectionStore.isMiniplayer).toBe(true);
-
-    await collectionStore.exitMiniplayerMode();
-    expect(invoke).toHaveBeenCalledWith("exit_miniplayer_mode", {});
-    expect(collectionStore.isMiniplayer).toBe(false);
-  });
-
-  it("captures a settled resize/move into the full player's geometry", async () => {
-    vi.useFakeTimers();
-    try {
-      expect(collectionStore.isMiniplayer).toBe(false);
-      vi.mocked(invoke).mockImplementation(async (cmd: string) => {
-        if (cmd === "get_window_geometry") return { width: 1500, height: 950, x: 10, y: 20 };
-        return null;
-      });
-
-      windowGeometryCallbacks.forEach((cb) => cb());
-      await vi.advanceTimersByTimeAsync(400);
-
-      expect(collectionStore.savedWindowWidth).toBe(1500);
-      expect(collectionStore.savedWindowHeight).toBe(950);
-      expect(collectionStore.savedWindowX).toBe(10);
-      expect(collectionStore.savedWindowY).toBe(20);
-    } finally {
-      vi.useRealTimers();
-    }
-  });
-
-  it("skips geometry capture while a mode toggle is in flight, then captures once settled", async () => {
-    // Regression guard for the growth/shrink-per-toggle bug: a resize/move
-    // event firing mid-transition (which includes the toggle's own
-    // programmatic resize/reposition) must not be captured as genuine.
-    vi.useFakeTimers();
-    try {
-      let resolveEnter: (value: unknown) => void = () => {};
-      const pendingEnter = new Promise((resolve) => {
-        resolveEnter = resolve;
-      });
-
-      vi.mocked(invoke).mockImplementation(async (cmd: string) => {
-        if (cmd === "geometry_capture_supported") return true;
-        if (cmd === "enter_miniplayer_mode") return pendingEnter;
-        if (cmd === "get_window_geometry") return { width: 1500, height: 950, x: null, y: null };
-        return null;
-      });
-
-      collectionStore.setMiniplayerGeometry(300, 360);
-      const enterPromise = collectionStore.enterMiniplayerMode();
-
-      windowGeometryCallbacks.forEach((cb) => cb());
-      await vi.advanceTimersByTimeAsync(400);
-      expect(collectionStore.miniplayerWidth).toBe(300);
-
-      resolveEnter(undefined);
-      await enterPromise;
-
-      windowGeometryCallbacks.forEach((cb) => cb());
-      await vi.advanceTimersByTimeAsync(400);
-      expect(collectionStore.miniplayerWidth).toBe(1500);
-
-      await collectionStore.exitMiniplayerMode();
-      expect(collectionStore.isMiniplayer).toBe(false);
-    } finally {
-      vi.useRealTimers();
-    }
-  });
-
-  it("blocks exitMiniplayerMode while a prior enterMiniplayerMode call is still in flight", async () => {
-    // Rapid back-and-forth toggling (e.g. holding Ctrl+M) could otherwise let
-    // exit start reading/resizing the window before enter's own resize IPC
-    // call has resolved, corrupting the captured "current size" baseline and
-    // compounding into runaway growth on repeated toggles.
-    let resolveEnter: (value: unknown) => void = () => {};
-    const pendingEnter = new Promise((resolve) => {
-      resolveEnter = resolve;
-    });
-    let exitCallCount = 0;
-
-    vi.mocked(invoke).mockImplementation(async (cmd: string) => {
-      if (cmd === "enter_miniplayer_mode") return pendingEnter;
-      if (cmd === "exit_miniplayer_mode") {
-        exitCallCount++;
-        return null;
-      }
-      return null;
-    });
-
-    const enterCall = collectionStore.enterMiniplayerMode();
-    expect(collectionStore.isMiniplayer).toBe(true);
-
-    await collectionStore.exitMiniplayerMode();
-    expect(exitCallCount).toBe(0);
-    expect(collectionStore.isMiniplayer).toBe(true);
-
-    resolveEnter(undefined);
-    await enterCall;
-
-    await collectionStore.exitMiniplayerMode();
-    expect(exitCallCount).toBe(1);
-    expect(collectionStore.isMiniplayer).toBe(false);
-  });
-
-  it("manages recent searches state, deduplication, and persistence", () => {
-    collectionStore.clearRecentSearches();
-    expect(collectionStore.recentSearches).toHaveLength(0);
-
-    collectionStore.addRecentSearch({
-      kind: "query",
-      title: "Pink Floyd",
-      query: "Pink Floyd"
-    });
-    expect(collectionStore.recentSearches).toHaveLength(1);
-    expect(collectionStore.recentSearches[0].title).toBe("Pink Floyd");
-
-    // Deduplication test
-    collectionStore.addRecentSearch({
-      kind: "artist",
-      title: "The Beatles",
-      subtitle: "Artist"
-    });
-    collectionStore.addRecentSearch({
-      kind: "query",
-      title: "Pink Floyd",
-      query: "Pink Floyd"
-    });
-    expect(collectionStore.recentSearches).toHaveLength(2);
-    expect(collectionStore.recentSearches[0].title).toBe("Pink Floyd");
-
-    // Capacity cap (max 10) test
-    for (let i = 1; i <= 12; i++) {
-      collectionStore.addRecentSearch({
-        kind: "query",
-        title: `Search ${i}`,
-        query: `Search ${i}`
-      });
-    }
-    expect(collectionStore.recentSearches).toHaveLength(10);
-    expect(collectionStore.recentSearches[0].title).toBe("Search 12");
-
-    // Remove single item
-    const itemToRemove = collectionStore.recentSearches[0];
-    collectionStore.removeRecentSearch(itemToRemove.id);
-    expect(collectionStore.recentSearches).toHaveLength(9);
-    expect(collectionStore.recentSearches.some(r => r.id === itemToRemove.id)).toBe(false);
-
-    // Clear all
-    collectionStore.clearRecentSearches();
-    expect(collectionStore.recentSearches).toHaveLength(0);
-    expect(localStorage.getItem("luminous_recent_searches")).toBe("[]");
-  });
-
-  it("restores miniplayer mode synchronously from localStorage on startup and syncs backend IPC", async () => {
-    localStorage.setItem("layout_isMiniplayer", "true");
-    vi.mocked(invoke).mockResolvedValue(null);
-
-    // Re-initialize collection store
-    await collectionStore.enterMiniplayerMode(true);
-    expect(collectionStore.isMiniplayer).toBe(true);
-    expect(invoke).toHaveBeenCalledWith("enter_miniplayer_mode", expect.any(Object));
-  });
-});
